requirements.txt
streamlit
pandas
seaborn
plotly
matplotlib
wordcloud
scikit-learn
numpy
requests
python-docx
PyPDF2
openpyxl

